Fix worker crash from bad socket client
This bug fixes a problem caused by a socket client quickly disconnecting
after requesting 'data' from a worker socket. I have so far been unable
to track down what has been causing this bad client behavior, but it
seems sensible that a socket client should not be able to crash a worker
like this, so I hope this gets merged.