fix warnings when values set in .json file minUptime and spinSleepTime
Warnings are thrown during getOptions if minUptime and/or spinSleepTime are not set as CLI arguments. If these parameters are defined in a .json config file, the program flow first checks for the CLI argument, throws  the warning and then sets the options based on the .json parameters. This is a simple fix to the if statement which also checks if these parameters are set in the .json file.