Add dir column by default on "forever list"
This adds the "dir" column by default when issuing the "forever list". It is the only field that can help the user to determine which PID is currently running but it is hidden by default!

Moreover adding an UID works fine however if the UID is added to the package.json then it will be commited in GIT thus not allowing the user to know if is a dev or a prod deployment.